modularize js files for home page
