+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "scripts": {
     "build": "node build.js",
-    "serve": "live-server docs --port=8080 --entry-file=index.html --mount=/GLTP:./docs --open=/GLTP --watch=src"
+    "serve": "live-server docs --port=8080 --entry-file=index.html --mount=/GLTP:./docs --open=/GLTP --watch=src --cors"
   },
   "devDependencies": {
     "live-server": "^1.2.2"
