+++ src/home.js
@@ -1,0 +1,41 @@
+import { dataUrl, setupNavigation } from './shared.js';
+import { MapsTable } from './maps.js';
+import { Leaderboard, processLeaderboardData } from './leaderboard.js';
+
+// Load presets
+const presets = await fetch(`./presets.json`)
+    .then(response => response.json());
+
+// Setup navigation
+setupNavigation();
+
+// Fetch and process data
+fetch(dataUrl)
+  .then(response => response.json())
+  .then(data => {
+    // Process data for both maps and leaderboards
+    const {
+      gamesCompletedLeaderboard,
+      worldRecordsLeaderboard,
+      soloWorldRecordsLeaderboard,
+      cappingWorldRecordsLeaderboard,
+      bestRecords,
+      recordsByMap
+    } = processLeaderboardData(data);
+
+    // Initialize and render maps table
+    const mapsTable = new MapsTable(presets, recordsByMap);
+    const recordsArray = Object.values(bestRecords);
+    recordsArray.sort((a, b) => new Date(b.timestamp) - new Date(a.timestamp));
+    mapsTable.render(recordsArray);
+
+    // Initialize and render leaderboards
+    const leaderboard = new Leaderboard();
+    leaderboard.render(
+      worldRecordsLeaderboard,
+      soloWorldRecordsLeaderboard,
+      cappingWorldRecordsLeaderboard,
+      gamesCompletedLeaderboard
+    );
+  })
+  .catch(error => console.error("Error fetching data:", error));
