+++ src/index.html
@@ -44,7 +44,7 @@
     <div id="leaderboardContainer"></div>
   </div>
 
-  <script type="module" src="script.js"></script>
+  <script type="module" src="home.js"></script>
 
 </body>
 </html>
